Move legacy-query-server/ to query-server/legacy/
