+++ extensions/ql-vscode/src/query-server/legacy/legacyRunner.ts
@@ -1,15 +1,19 @@
 import { CancellationToken } from "vscode";
-import { CodeQLCliServer } from "../cli";
-import { ProgressCallback } from "../progress";
-import { Logger } from "../common";
-import { DatabaseItem } from "../local-databases";
+import { CodeQLCliServer } from "../../cli";
+import { ProgressCallback } from "../../progress";
+import { Logger } from "../../common";
+import { DatabaseItem } from "../../local-databases";
 import {
   Dataset,
   deregisterDatabases,
   registerDatabases,
-} from "../pure/legacy-messages";
-import { CoreQueryResults, CoreQueryTarget, QueryRunner } from "../queryRunner";
-import { QueryOutputDir } from "../run-queries-shared";
+} from "../../pure/legacy-messages";
+import {
+  CoreQueryResults,
+  CoreQueryTarget,
+  QueryRunner,
+} from "../../queryRunner";
+import { QueryOutputDir } from "../../run-queries-shared";
 import { QueryServerClient } from "./queryserver-client";
 import {
   clearCacheInDatabase,
