+++ extensions/ql-vscode/test/vscode-tests/cli-integration/legacy-query.test.ts
@@ -4,11 +4,11 @@ import { dirSync } from "tmp";
 import { pathToFileURL } from "url";
 import { CancellationTokenSource } from "vscode-jsonrpc";
 import * as messages from "../../../src/pure/legacy-messages";
-import * as qsClient from "../../../src/legacy-query-server/queryserver-client";
+import * as qsClient from "../../../src/query-server/legacy/queryserver-client";
 import * as cli from "../../../src/cli";
 import { CellValue } from "../../../src/pure/bqrs-cli-types";
 import { describeWithCodeQL } from "../cli";
-import { QueryServerClient } from "../../../src/legacy-query-server/queryserver-client";
+import { QueryServerClient } from "../../../src/query-server/legacy/queryserver-client";
 import { extLogger, ProgressReporter } from "../../../src/common";
 import { createMockApp } from "../../__mocks__/appMock";
 import { getActivatedExtension } from "../global.helper";
